Temporarily remove AddressPolicy from the public API
I would like to do more work on the API and implementation
before committing to this API, and I want to do that after
we cut 5.0 final.
