+++ okhttp/api/jvm/okhttp.api
@@ -339,16 +339,6 @@ public final class okhttp3/ConnectionPool {
 	public final fun connectionCount ()I
 	public final fun evictAll ()V
 	public final fun idleConnectionCount ()I
-	public final fun setPolicy (Lokhttp3/Address;Lokhttp3/ConnectionPool$AddressPolicy;)V
-}
-
-public final class okhttp3/ConnectionPool$AddressPolicy {
-	public final field backoffDelayMillis J
-	public final field backoffJitterMillis I
-	public final field minimumConcurrentCalls I
-	public fun <init> ()V
-	public fun <init> (IJI)V
-	public synthetic fun <init> (IJIILkotlin/jvm/internal/DefaultConstructorMarker;)V
 }
 
 public final class okhttp3/ConnectionSpec {
